:art: refactor: 상품 도메인 모듈 분리
