+++ core/core-product-api/src/main/resources/core-product-api.yml
