+++ core/core-product-domain/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    compileOnly 'org.projectlombok:lombok'
+    annotationProcessor 'org.projectlombok:lombok'
+}
