+++ core/core-product-domain/src/main/java/com/phoenix/logistics/core/product/domain/model/Product.java
