+++ core/core-product-domain/src/main/java/com/phoenix/logistics/core/product/domain/repository/ProductRepository.java
@@ -1,10 +1,7 @@
 package com.phoenix.logistics.core.product.domain.repository;
 
-import org.springframework.data.repository.NoRepositoryBean;
-
 import com.phoenix.logistics.core.product.domain.model.Product;
 
-@NoRepositoryBean
 public interface ProductRepository {
 
     Product save(Product product);
