+++ core/core-product-domain/src/main/resources/core-product-domain.yml
@@ -1,0 +1,3 @@
+spring:
+  application:
+    name: core-product-domain
