+++ settings.gradle
@@ -15,6 +15,7 @@ include 'core:core-user'
 include 'core:core-product-api'
 include 'core:core-hub-api'
 include 'core:core-hub-domain'
+include 'core:core-product-domain'
 include 'storage:db-core'
 include 'storage:db-core-product'
 include 'storage:db-core-user'
