+++ storage/db-core-product/build.gradle
@@ -1,5 +1,7 @@
 dependencies {
-    api 'org.springframework.boot:spring-boot-starter-data-jpa'
+    compileOnly project(":core:core-product-domain")
+
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'com.h2database:h2'
 
     compileOnly 'org.projectlombok:lombok'
